songplayer

+++ MelodiUS/include/generator.h
@@ -14,7 +14,7 @@ constexpr double SINE_AMPLITUDE = 0.8;
 /*****************************************************************************/
 /* Function declarations --------------------------------------------------- */
 [[nodiscard]] Recording Generate_Sine(size_t freq,
-                                      size_t numSeconds      = NUM_SECONDS,
+                                      double numSeconds      = NUM_SECONDS,
                                       size_t sampleRate      = SAMPLE_RATE,
                                       size_t numChannels     = NUM_CHANNELS,
                                       size_t framesPerBuffer = FRAMES_PER_BUFFER,
